파일분리 및 정리 (20.11.19)
파일분리 안된거 이동 및 정리
